refactor: 플레이어 파싱 테스트 추가
